目前两个重大bug： 1. 拍照后照片无法存储 2. CSV读入 未完成内容： 1. Upload数据读取合再分发

+++ app/src/main/java/com/example/icst/StudentActivity.java
@@ -1,6 +1,8 @@
 package com.example.icst;
 
+import android.Manifest;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.graphics.drawable.BitmapDrawable;
@@ -10,6 +12,8 @@ import android.os.Environment;
 import android.provider.MediaStore;
 import android.support.design.widget.CollapsingToolbarLayout;
 import android.support.design.widget.FloatingActionButton;
+import android.support.v4.content.ContextCompat;
+import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.util.Log;
@@ -138,31 +142,43 @@ public class StudentActivity extends AppCompatActivity {
         fab.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                String filePath = "Student" + Long.toString(id) + ".jpg";
-                if (Environment.MEDIA_MOUNTED.equals(Environment.getExternalStorageState())) {
-                    File dir = getExternalFilesDir(Environment.DIRECTORY_PICTURES);
-                    File outputImage = new File(dir, filePath);
-                    try {
-                        if (outputImage.exists()) {
-                            outputImage.delete();
+                if (ContextCompat.checkSelfPermission(StudentActivity.this, Manifest.permission.READ_SMS)
+                        == PackageManager.PERMISSION_GRANTED &&
+                        ContextCompat.checkSelfPermission(StudentActivity.this, Manifest.permission.WRITE_EXTERNAL_STORAGE)
+                                == PackageManager.PERMISSION_GRANTED) {
+                    String filePath = "Student" + Long.toString(id) + ".jpg";
+                    if (Environment.MEDIA_MOUNTED.equals(Environment.getExternalStorageState())) {
+                        File dir = getExternalFilesDir(Environment.DIRECTORY_PICTURES);
+                        File outputImage = new File(dir, filePath);
+                        try {
+                            if (outputImage.exists()) {
+                                outputImage.delete();
+                            }
+                            outputImage.createNewFile();
+                        } catch (IOException e) {
+                            e.printStackTrace();
                         }
-                        outputImage.createNewFile();
-                    } catch (IOException e) {
-                        e.printStackTrace();
-                    }
-                    //将File对象转换为Uri并启动照相程序
+                        //将File对象转换为Uri并启动照相程序
 
-                    Intent intent = new Intent();
-                    // 指定开启系统相机的Action
-                    intent.setAction(MediaStore.ACTION_IMAGE_CAPTURE);
-                    intent.addCategory(Intent.CATEGORY_DEFAULT);
-                    // 把文件地址转换成Uri格式
-                    Uri uri = Uri.fromFile(outputImage);
-                    // 设置系统相机拍摄照片完成后图片文件的存放地址
-                    intent.putExtra(MediaStore.EXTRA_OUTPUT, uri);
-                    intent.putExtra("filePath", filePath);
-                    startActivityForResult(intent, 0);
-                    //拍完照startActivityForResult() 结果返回onActivityResult()函数
+                        Intent intent = new Intent();
+                        // 指定开启系统相机的Action
+                        intent.setAction(MediaStore.ACTION_IMAGE_CAPTURE);
+                        intent.addCategory(Intent.CATEGORY_DEFAULT);
+                        // 把文件地址转换成Uri格式
+                        Uri uri = Uri.fromFile(outputImage);
+                        // 设置系统相机拍摄照片完成后图片文件的存放地址
+                        intent.putExtra(MediaStore.EXTRA_OUTPUT, uri);
+                        intent.putExtra("filePath", filePath);
+                        startActivityForResult(intent, 0);
+                        //拍完照startActivityForResult() 结果返回onActivityResult()函数
+                    }
+                } else {
+                    new AlertDialog.Builder(StudentActivity.this)
+                            .setTitle("缺少权限")
+                            .setIcon(R.drawable.ic_warning)
+                            .setMessage("需要相机和存储权限")
+                            .setPositiveButton("确定", null)
+                            .show();
                 }
             }
         });
